init_dyn_array: remove type-specific code
Add code to initialize array with help of type param. This allows the
array to resize without any type-specific code--the size of the member
is stored on initialization and used internally. Additionally, convert
type name to its size via a new interim macro for better readability.

This will allow us to consolidate *_node and *_char functions into a few
general-purpose functions.